requirements.txt
aiomysql==0.0.20
amqp==2.5.2
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextvars==2.4
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.1
fastapi==0.54.2
h11==0.9.0
httptools==0.1.1
idna==2.9
immutables==0.14
importlib-metadata==1.7.0
kombu==4.6.8
mock==4.0.2
more-itertools==8.3.0
packaging==20.4
passlib==1.7.2
peewee==3.13.3
peewee-async==0.7.0
peewee-migrations==0.3.18
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
PyMySQL==0.9.2
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
redis==3.2.1
requests==2.23.0
six==1.14.0
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.8
uvloop==0.14.0
vine==1.3.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0